requirements.txt
python-dotenv
streamlit
transformers

#pytorch
torch
torchvision
torchaudio

#llm
langchain
langchain-community
langchain-google-genai
google.generativeai
openai
faiss-gpu

#vector database
astrapy
cassio
datasets
tiktoken


#notebook
ipykernel

#pdf
pypdf

#voice
soundfile

#image
Pillow

#embedding
BCEmbedding==0.1.1
